+++ R/availableCores.R
@@ -35,6 +35,10 @@
 #'
 #' @param omit (integer; non-negative) Number of cores to not include.
 #'
+#' @param max (integer; positive) Maximum number of cores returned.
+#' `availableCores(..., max = n)` is short for
+#' `min(n, availableCores(...), na.rm = TRUE)`.
+#'
 #' @return Return a positive (>= 1) integer.
 #' If `which = "all"`, then more than one value may be returned.
 #' Together with `na.rm = FALSE` missing values may also be returned.
@@ -231,7 +235,7 @@
 #'
 #' @importFrom parallel detectCores
 #' @export
-availableCores <- function(constraints = NULL, methods = getOption2("parallelly.availableCores.methods", c("system", "/proc/self/status", "cgroups.cpuset", "cgroups.cpuquota", "cgroups2.cpu.max", "nproc", "mc.cores", "BiocParallel", "_R_CHECK_LIMIT_CORES_", "Bioconductor", "LSF", "PJM", "PBS", "SGE", "Slurm", "fallback", "custom")), na.rm = TRUE, logical = getOption2("parallelly.availableCores.logical", TRUE), default = c(current = 1L), which = c("min", "max", "all"), omit = getOption2("parallelly.availableCores.omit", 0L)) {
+availableCores <- function(constraints = NULL, methods = getOption2("parallelly.availableCores.methods", c("system", "/proc/self/status", "cgroups.cpuset", "cgroups.cpuquota", "cgroups2.cpu.max", "nproc", "mc.cores", "BiocParallel", "_R_CHECK_LIMIT_CORES_", "Bioconductor", "LSF", "PJM", "PBS", "SGE", "Slurm", "fallback", "custom")), na.rm = TRUE, logical = getOption2("parallelly.availableCores.logical", TRUE), default = c(current = 1L), which = c("min", "max", "all"), omit = getOption2("parallelly.availableCores.omit", 0L), max = getOption2("parallelly.availableCores.max", Inf)) {
   stop_if_not(
     is.null(constraints) || is.character(constraints), !anyNA(constraints)
   )
@@ -248,7 +252,9 @@ availableCores <- function(constraints = NULL, methods = getOption2("parallelly.
   stop_if_not(length(omit) == 1L, is.numeric(omit),
               is.finite(omit), omit >= 0L)
   omit <- as.integer(omit)
-  
+
+  stop_if_not(length(max) == 1L, is.numeric(max), !is.na(max), max >= 1L)
+
   ncores <- rep(NA_integer_, times = length(methods))
   names(ncores) <- methods
   for (kk in seq_along(methods)) {
@@ -452,6 +458,11 @@ availableCores <- function(constraints = NULL, methods = getOption2("parallelly.
   if (omit > 0L) {
     ncores <- ncores - omit
     ncores[ncores < 1L] <- 1L
+  }
+
+  ## Upper limit?
+  if (is.finite(max)) {
+    ncores <- min(max, ncores, na.rm = TRUE)
   }
 
   ## Sanity check